Introduction

WorkshopWizard.jl is awesome.

Index

WorkshopWizard.check_IJulia

WorkshopWizard.download

WorkshopWizard.install

WorkshopWizard.install_IJulia

WorkshopWizard.run_wizard

Functions

WorkshopWizard.run_wizard — Method

run_wizard()

Start the workshop installation wizard. It will interactively guide you through the selection, download, and installation of a Julia workshop.

If not already present, the wizard will also install IJulia.

WorkshopWizard.check_IJulia — Method

Check if IJulia is available.

WorkshopWizard.install_IJulia — Method

install_IJulia(; python = nothing, globally = true)

Install the IJulia kernel for jupyter.

Although the installer should locate your local python installation automatically (if present), the keyword python can be used to manually point to a python executable.

When globally = true IJulia will be installed to the global v#.# environment, or to the environment that has been given to --project when Julia was started.

WorkshopWizard.download — Method

download(; repo = default_repo(), path = default_path())

Download a workshop from github.

Keyword arguments repo and path can be used to specify the github repository and the local installation directory.

WorkshopWizard.install — Method

install(; repo = default_repo(), path = default_path(), check_IJulia = true, auto_overwrite = true)

More programmatic workshop installation interface.

By default, the workshop will be downloaded to the desktop (on windows) or the home directory (on linux/macOS). The installation path can be adjusted per keyword argument path = desired/install/path.

Missing docstring.

Missing docstring for default_path(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for default_repo(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for default_workshop(). Check Documenter's build log for details.