Functions

Index

WorkshopWizard.download

WorkshopWizard.install

WorkshopWizard.install_IJulia

WorkshopWizard.install_dependencies

WorkshopWizard.run_wizard

WorkshopWizard.uninstall_IJulia

List of Functions

WorkshopWizard.run_wizard — Function

run_wizard()

Start the workshop installation wizard. It will interactively guide you through the selection, download, and installation of a Julia workshop.

If not already present, the wizard will also install IJulia.

WorkshopWizard.install_IJulia — Function

install_IJulia(; python = nothing)

Install IJulia to the to the global v#.# environment, or to the environment that has been given to --project when Julia was started.

IJulia should automatically install python and jupyter if necessary. The keyword python can be used to manually point to a python executable (to give it a hint).

WorkshopWizard.uninstall_IJulia — Function

uninstall_IJulia()

Remove IJulia from the global environment.

WorkshopWizard.download — Function

download(; repo = default_repo(), path = default_path())

Download a workshop from github.

Keyword arguments repo and path can be used to specify the github repository and the local installation directory.

WorkshopWizard.install_dependencies — Function

install_dependencies(workshop_path)

Instantiates and precompiles the workshops Julia environment (Manifest.toml). Depending on the workshop and your machine this may take a while (even >15 minutes).

WorkshopWizard.install — Function

install(; repo = default_repo(), path = default_path(), check_IJulia = true, auto_overwrite = true)

More programmatic workshop installation interface.

By default, the workshop will be downloaded to the desktop (on windows) or the home directory (on linux/macOS). The installation path can be adjusted per keyword argument path = desired/install/path.